module5-tast1: удалил function.js добавил новые файлы js
