+++ index.html
@@ -231,7 +231,9 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="./js/function.js"></script>
-  <script src="./js/main.js"></script>
+  <script type="module" src="./js/main.js"></script>
+  <script type="module" src="./js/data.js"></script>
+  <script type="module" src="./js/until.js"></script>
+  <script type="module" src="./js/gallery.js"></script>
 </body>
 </html>
